Simplifies the implementation, we only need to have UmbracoTreeSearcher
