+++ src/Umbraco.Web/Runtime/WebInitialComposer.cs
@@ -148,7 +148,6 @@ namespace Umbraco.Web.Runtime
                 .Add(() => composition.TypeLoader.GetTypes<ISearchableTree>());
 
             composition.Register<UmbracoTreeSearcher>(Lifetime.Request);
-            composition.RegisterUnique<UmbracoEntitySearcher>();
 
             composition.WithCollectionBuilder<EditorValidatorCollectionBuilder>()
                 .Add(() => composition.TypeLoader.GetTypes<IEditorValidator>());
