+++ src/Umbraco.Web/Trees/ContentTreeControllerBase.cs
@@ -18,11 +18,23 @@ using System.Web.Http.ModelBinding;
 using Umbraco.Web.Actions;
 using Umbraco.Web.Composing;
 using Umbraco.Core.Security;
+using Umbraco.Core.Cache;
+using Umbraco.Core.Configuration;
+using Umbraco.Core.Persistence;
 
 namespace Umbraco.Web.Trees
 {
     public abstract class ContentTreeControllerBase : TreeController
     {
+
+        protected ContentTreeControllerBase(IGlobalSettings globalSettings, IUmbracoContextAccessor umbracoContextAccessor, ISqlContext sqlContext, ServiceContext services, AppCaches appCaches, IProfilingLogger logger, IRuntimeState runtimeState, UmbracoHelper umbracoHelper) : base(globalSettings, umbracoContextAccessor, sqlContext, services, appCaches, logger, runtimeState, umbracoHelper)
+        {
+        }
+
+        protected ContentTreeControllerBase()
+        {
+        }
+
         #region Actions
 
         /// <summary>
@@ -528,6 +540,8 @@ namespace Umbraco.Web.Trees
         private readonly ConcurrentDictionary<string, IEntitySlim> _entityCache = new ConcurrentDictionary<string, IEntitySlim>();
 
         private bool? _ignoreUserStartNodes;
+
+        
 
         /// <summary>
         /// If the request should allows a user to choose nodes that they normally don't have access to
